Update tox.ini to latest settings
This updates the minversion used in tox.ini to be a more recent version
to match other openstack project's settings. It also drops the out of
date defaults and replaces them with venv since there are not actually
any unit test environments to run. The only useful environment is venv,
so the default target will now just create the venv as a shortcut.

Change-Id: I416ed1cc997cd49029cf8fec31b30df5f4179ff2
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>